+++ src/contrib/loggers/Akka.Logger.Serilog/Event/Serilog/SerilogLogMessageFormatter.cs
@@ -1,37 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Akka.Event;
-using Serilog.Events;
-using Serilog.Parsing;
+using System.Text;
+using System.Threading.Tasks;
 
+// ReSharper disable once CheckNamespace
 namespace Akka.Serilog.Event.Serilog
-{
-    public class SerilogLogMessageFormatter : ILogMessageFormatter
-    {
-        private readonly MessageTemplateCache _templateCache;
-            
-        public SerilogLogMessageFormatter()
-        {
-            _templateCache = new MessageTemplateCache(new MessageTemplateParser());
-        }
-
-        public string Format(string format, params object[] args)
-        {
-            var template = _templateCache.Parse(format);
-            var propertyTokens = template.Tokens.OfType<PropertyToken>().ToArray();
-            var properties = new Dictionary<string, LogEventPropertyValue>();
-
-            for (var i = 0; i < args.Length; i++)
-            {
-                var propertyToken = propertyTokens.ElementAtOrDefault(i);
-                if (propertyToken == null)
-                    break;
-
-                properties.Add(propertyToken.PropertyName, new ScalarValue(args[i]));
-            }
-
-            return template.Render(properties);
-        }
-    }
+{    
+	[Obsolete("Use Akka.Logger.Serilog.SerilogLogMessageFormatter instead. This class will be removed in a future version.")]
+	public class SerilogLogMessageFormatter : global::Akka.Logger.Serilog.SerilogLogMessageFormatter
+	{
+	}
 }
